backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg    
alembic
pydantic
pydantic-settings
python-dotenv
pytest
pytest-asyncio
httpx
python-dateutil
